Update INSTALL to reference Qt6 and the new way of building this package.
In commit 67dad47ba, sysmon-qt.pro was modified to have both 'src' and
'data' as subdirectories. This was done to allow the package to be
installed via 'make install'. Because of this though, we no longer need
to change directory into 'src', and the instructions are no longer
correct. We could probably use "qmake6 ../sysmon-qt.pro" from within the
src subdirectory, but let's do it this way since it's cleaner.

In addition, change the URL for Qt to reference Qt6 instead of
Qt5/Qt5-alternate, and a fix a typo in the NEWS file